First step to add Only Sexy comments feature.
I find this plugin very useful but actually the only feature I use
most of the time is the so called "Sexy Comments", however this feature doesn't
behaves exactly as I want. For example:

Default behavior for toggle comments:

    //while (true) {
        //Socket client = sock.accept();
        //PrintWriter pout = new PrintWriter(client.getOutputStream(), true);
        //pout.println(new java.util.Date().toString());
        //client.close();
    //}

Default behavior for sexy comments:
    /*
     *while (true) {
     *    Socket client = sock.accept();
     *    PrintWriter pout = new PrintWriter(client.getOutputStream(), true);
     *    pout.println(new java.util.Date().toString());
     *    client.close();
     *}
     */

Desired behavior in both cases:

    //while (true) {
    //    Socket client = sock.accept();
    //    PrintWriter pout = new PrintWriter(client.getOutputStream(), true);
    //    pout.println(new java.util.Date().toString());
    //    client.close();
    //}

The feature I'm proposing is to create a mode that, when ON, the Toggle function calls
CommentLinesSexy() by default and this function behaves as shown above.

There is a lot of work to do for this feature to work with all the languages
supported, this is only the first step. The tested languages are:

- Python
- C/C++
- Java
- Ruby
- Bash
- VimML

Personally I have disabled all the mapping of this plugin and only have
<leader>c binded to SexyCommentsOnly. Don't worry this isn't part of my
SexyCommentsOnly feature.