Remove apparently bogus filetype
I can't find any record of this being a legitimate filetype. It isn't in
vim or nvim's filetype detection system and nothing comes up in search.
If it's legit somebody can complain and we'll add it back, but it seems
likely this was left over from a fat-fingered mistake during the initial
development of the plugin.

Fixes #363