Use HTML comment syntax for Rmarkdown
Closes #415

While Markdown doesn't officially have a comment syntax, HTML comments
are widely parsed as such. This is what most Markdown treat as Markdown
comments. As a Markdown derivative it seams logical to do the same thing
in Rmarkdown.

The previously added '#' syntax seems to be for inside code blocks, and
that would be language dependent anyway. I've left the hash option as
the alternate.