Remove obsolete file comments header
The days when vim plugins were distributed as individual files are long
gone, these days they are almost always git repositories or tarballs. As
such the relevant meta data is included, just not in this way. Not only
does this information have a habit of being absurdly out of date, it's
not really useful any more to try to update it.