Remove contributor list from help docs, link to Contributors
Not only is the list hopelessly out of date, it's almost impossible to
keep it up to date in any balance way. The only reasonably thing I could
think to put here would be a normalized list of authors scrapped from
git commit history, but Github's view of that is nicer anyway.